fix: credit card info
